# openapi/v1.yml
openapi: '3.1.0'
info:
  title: TwextHub API
  version: '1.0.0'
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    TwextHub is a lightweight registry for publishing and discovering
    Twext-compiled TurboWarp extensions ("Twexts"). Packages are namespaced
    to their publishing account (`@namespace/id`) and versioned with SemVer.
    Publishing sends a gzipped tarball of the Twext project (`twext.yml` plus
    sources); the hub validates the manifest and compiles the extension
    itself in a sandboxed build, so the reviewed source is exactly what
    produces the served blob.

    The `/v1` surface documented here is stable as of 1.0.0. Breaking changes
    require a new major version; additive changes bump the minor (see
    CHANGELOG.md).

    Every authenticated (`Userspace`/`Adminspace`) endpoint requires the
    caller to have accepted the current Terms of Service; a caller who
    hasn't gets a `403` (see `Forbidden`) until they patch their own user with
    `termsAcceptedVersion` set to the `version` that `GET /terms` reports. A
    stale version is a `422`, so an acceptance is always recorded against text
    the account was shown. Exemptions that stay available without
    re-acceptance: `DELETE /sessions/current`, `GET /me`, the acceptance
    itself, password-only user updates, and account deletion.

    Bearer tokens come in two kinds. **Sessions** are created by
...
    every scope. **Automation tokens** are created explicitly via
...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /sessions/current:
    delete:
      summary: Revoke the session making this request
      operationId: revokeCurrentSession
      tags: [Userspace]
      description: >
        `current` is the session that authenticated the request, so a client
        that never kept its own id can still sign itself out. Exempt from the
        terms gate: an account that has just been bumped to new terms is exactly
        the one that needs to be able to end its session. Session-authenticated
        only, with no scope that would let a token in: the route deletes by the
        caller's own token id and the `sessions` and `automation_tokens` id
        sequences are independent, so a token reaching it could end another
        account's session.
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /tokens:
    get:
      summary: List a caller's automation tokens
      operationId: listAutomationTokens
      tags: [Userspace]
      description: Needs the `manage:tokens` scope. `namespace` defaults to the caller; an admin may pass it to inspect another account's tokens. Never returns token secrets.
      parameters:
        - name: namespace
          in: query
          schema: { $ref: '#/components/schemas/Namespace' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Dir'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated automation tokens
          content:
            application/json:
              schema:
                type: object
                required: [data, _links]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AutomationToken' }
                  _links: { $ref: '#/components/schemas/PageLinks' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      summary: Create an automation token
      operationId: createAutomationToken
      tags: [Userspace]
      description: Needs the `manage:tokens` scope. Every scope requested must be one the caller already holds. The `token` secret is returned exactly once, in this response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string, maxLength: 80, description: 'e.g. "CI publish".' }
                scopes:
                  type: array
                  items: { $ref: '#/components/schemas/Scope' }
                  minItems: 1
                expiresInDays:
                  { type: integer, minimum: 1, description: Omit for a token that never expires. }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationTokenCreated' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /tokens/current:
    delete:
      summary: Revoke the automation token making this request
      operationId: revokeCurrentAutomationToken
      tags: [Userspace]
      description: >
        The automation-token counterpart to `DELETE /sessions/current`, for CI
        that holds a single token and no id. Automation-token authenticated, so a
        session gets a `403`: it has no token of its own to revoke. This route
        needs no scope, since revoking the credential in hand is something any
        token may do for itself.
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tokens/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Rename or rescope an automation token
      operationId: updateAutomationToken
      tags: [Userspace]
      description: Self or admin, needs the `manage:tokens` scope. New scopes must be ones the caller already holds. Never returns or reissues the token secret.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name: { type: string, maxLength: 80 }
                scopes:
                  type: array
                  items: { $ref: '#/components/schemas/Scope' }
                  minItems: 1
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationToken' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      summary: Revoke an automation token
      operationId: deleteAutomationToken
      tags: [Userspace]
      description: Self or admin, needs the `manage:tokens` scope.
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------------- Users
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: >
        Opaque position from a previous page, taken from that page's
        `_links.next` or `_links.prev` rather than assembled by hand.
    Dir:
      name: dir
      in: query
      schema: { type: string, enum: [back] }
      description: >
        Set to `back` to page towards the rows before the cursor. The server only
        looks for a cursor from a list sorted the other way round; anything else
        is a `400`. A client never needs it, because every page's `_links.prev`
        already sets it.
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
      description: Values above the server cap are clamped to the cap, not rejected. The default cap is 50.
...
  schemas:
    Namespace:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
...
    Scope:
      type: string
      enum:
        [
          publish,
          yank,
          read:source,
          manage:account,
          manage:orgs,
          manage:sessions,
          manage:tokens,
          admin,
        ]
      description: >-
        Only meaningful for automation tokens — sessions implicitly satisfy
        every scope. `publish` and `yank` cover the extension publish and
        unpublish routes; `read:source` covers fetching a version's source;
        `manage:account`, `manage:orgs`, `manage:sessions` and
        `manage:tokens` cover the settings, organization and credential
        routes a session would otherwise be the only caller of; `admin`
        covers `/admin/*` and version review, and is required in addition to
        the admin role rather than instead of it. A credential may only be
        granted a scope it already holds.
...
    AutomationToken:
      type: object
      required: [id, name, scopes, createdAt, expiresAt, lastUsedAt]
      properties:
        id: { type: string }
        name: { type: string }
        scopes:
          type: array
          items: { $ref: '#/components/schemas/Scope' }
        createdAt: { type: string, format: date-time }
        expiresAt:
          {
            type: ['string', 'null'],
            format: date-time,
            description: '`null` if the token never expires.',
          }
        lastUsedAt: { type: ['string', 'null'], format: date-time }
    AutomationTokenCreated:
      allOf:
        - $ref: '#/components/schemas/AutomationToken'
        - type: object
          required: [token]
          properties:
            token:
              {
                type: string,
                description: The bearer token secret. Shown only in this one response.,
              }
...
    PageLinks:
      type: object
      description: >
        The pages either side of this one, as complete URLs. Each already carries
        the cursor, the limit and the filters it was built from, so a client
        follows a link rather than assembling a query; a null means that
        direction is the end of the list.
      required: [self, next, prev]
      properties:
        self:
          type: string
          description: The request as it was made.
        next: { type: ['string', 'null'] }
        prev: { type: ['string', 'null'] }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: 'about:blank' }
        title: { type: string, example: 'Validation Error' }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Missing or invalid bearer token
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: >
        Authenticated, but not allowed to perform this action — e.g. not the
        resource owner or an admin, a namespace mismatch on publish, the
        current Terms of Service haven't been accepted yet
        (see `termsAcceptedVersion` on `PATCH /users/{namespace}`), or a
        credential missing the scope the route needs. The `detail` names the
        scope. An automation token on `/admin/*` or a version review also needs
        the `admin` scope on top of the account being an admin.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Resource not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    UnprocessableEntity:
      description: Semantically invalid input (e.g. non-SemVer version, id mismatch)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer